@@ -9,12 +9,12 @@
     "build-prod": "./node_modules/.bin/dojo build"
   },
   "dependencies": {
-    "@dojo/framework": "8.0.0-alpha.1"
+    "@dojo/framework": "8.0.0"
   },
   "devDependencies": {
-    "@dojo/cli": "^7.0.0",
-    "@dojo/cli-build-app": "^7.0.0",
-    "typescript": "3.5.2"
+    "@dojo/cli": "8.0.0",
+    "@dojo/cli-build-app": "8.0.0",
+    "typescript": "4.7.4"
   },
   "prettier": {
     "singleQuote": true,
