@@ -1,6 +1,6 @@
 {
   "name": "js-framework-benchmark-ef-js",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "ef.js demo",
   "main": "dist/main.js",
   "js-framework-benchmark": {
@@ -8,8 +8,9 @@
     "issues": []
   },
   "scripts": {
-    "build-dev": "rollup -c ./config/rollup.dev.js -w",
-    "build-prod": "cross-env NODE_ENV=production rollup -c ./config/rollup.prod.js"
+    "build": "rollup -c ./config/rollup.prod.mjs",
+    "build-dev": "rollup -c ./config/rollup.dev.mjs -w",
+    "build-prod": "cross-env NODE_ENV=production rollup -c ./config/rollup.prod.mjs"
   },
   "keywords": [
     "ef.js",
@@ -24,36 +25,36 @@
   },
   "devDependencies": {
     "@rollup/plugin-buble": "^0.21.3",
-    "@rollup/plugin-commonjs": "^19.0.0",
-    "@rollup/plugin-inject": "^4.0.2",
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "@rollup/plugin-inject": "^4.0.4",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@rollup/plugin-replace": "^2.4.2",
-    "camelcase": "^6.2.0",
-    "chalk": "^4.1.1",
-    "chokidar": "^3.5.2",
-    "connect-history-api-fallback": "^1.6.0",
+    "@rollup/plugin-node-resolve": "^14.1.0",
+    "@rollup/plugin-replace": "^4.0.0",
+    "camelcase": "^7.0.0",
+    "chalk": "^5.0.1",
+    "chokidar": "^3.5.3",
+    "connect-history-api-fallback": "^2.0.0",
     "cross-env": "^7.0.3",
-    "cssnano": "^5.0.6",
-    "del": "^6.0.0",
-    "del-cli": "^4.0.0",
-    "eft-parser": "^0.13.5",
-    "eslint": "^7.29.0",
-    "eslint-plugin-react": "^7.24.0",
-    "postcss": "^8.3.5",
+    "cssnano": "^5.1.13",
+    "del": "^7.0.0",
+    "del-cli": "^5.0.0",
+    "eft-parser": "^0.14.1",
+    "eslint": "^8.23.1",
+    "eslint-plugin-react": "^7.31.8",
+    "postcss": "^8.4.16",
     "postcss-nested": "5",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-simple-vars": "6",
-    "rollup": "^2.52.4",
+    "postcss-preset-env": "^7.8.2",
+    "postcss-simple-vars": "7",
+    "rollup": "^2.79.1",
     "rollup-plugin-browsersync": "^1.3.3",
-    "rollup-plugin-copy-glob": "^0.3.1",
-    "rollup-plugin-eft": "^0.13.10",
+    "rollup-plugin-copy-glob": "^0.3.2",
+    "rollup-plugin-eft": "^0.14.3",
     "rollup-plugin-eslint": "^7.0.0",
-    "rollup-plugin-postcss": "^4.0.0",
+    "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-progress": "^1.1.2",
     "rollup-plugin-terser": "^7.0.2"
   },
   "dependencies": {
-    "ef-core": "^0.13.11"
+    "ef-core": "^0.15.5"
   }
 }
