@@ -1,6 +1,6 @@
 {
   "name": "js-framework-benchmark-react-recoil",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "React Recoil demo",
   "main": "index.js",
   "js-framework-benchmark": {
@@ -31,8 +31,8 @@
     "webpack-cli": "3.3.4"
   },
   "dependencies": {
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "recoil": "0.2.0"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "recoil": "0.7.5"
   }
 }
