@@ -32,9 +32,9 @@
     "webpack-cli": "3.3.4"
   },
   "dependencies": {
-    "@react-rxjs/core": "0.6.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "rxjs": "6.6.3"
+    "@react-rxjs/core": "0.9.8",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "rxjs": "7.5.6"
   }
 }
