@@ -31,8 +31,8 @@
     "webpack-cli": "3.3.4"
   },
   "dependencies": {
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-tagged-state": "1.4.3"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-tagged-state": "1.21.1"
   }
 }
