@@ -1,6 +1,6 @@
 {
   "name": "js-framework-benchmark-react-zustand",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "React Zustand demo",
   "main": "index.js",
   "js-framework-benchmark": {
@@ -31,8 +31,8 @@
     "webpack-cli": "3.3.4"
   },
   "dependencies": {
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "zustand": "^3.6.7"
+    "react": "~18.2.0",
+    "react-dom": "~18.2.0",
+    "zustand": "4.1.1"
   }
 }
