@@ -25,6 +25,6 @@
     "rollup": "2.45.2",
     "rollup-plugin-svelte": "7.1.0",
     "rollup-plugin-terser": "7.0.2",
-    "svelte": "3.46.2"
+    "svelte": "3.50.1"
   }
 }
