@@ -1,5 +1,5 @@
 import { memo, useReducer } from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 const random = (max) => Math.round(Math.random() * 1000) % max;
 
@@ -121,4 +121,4 @@ const Main = () => {
   </div>);
 }
 
-render(<Main />, document.getElementById('main'));
+createRoot(document.getElementById("main")).render(<Main/>);
