@@ -1,63 +1,15 @@
-import React, { Component } from "react";
-import { createRoot } from "react-dom";
+import React from 'react';
+import { memo, useReducer } from 'react';
+import { createRoot } from 'react-dom';
 
 const random = (max) => Math.round(Math.random() * 1000) % max;
 
-const A = [
-  "pretty",
-  "large",
-  "big",
-  "small",
-  "tall",
-  "short",
-  "long",
-  "handsome",
-  "plain",
-  "quaint",
-  "clean",
-  "elegant",
-  "easy",
-  "angry",
-  "crazy",
-  "helpful",
-  "mushy",
-  "odd",
-  "unsightly",
-  "adorable",
-  "important",
-  "inexpensive",
-  "cheap",
-  "expensive",
-  "fancy",
-];
-const C = [
-  "red",
-  "yellow",
-  "blue",
-  "green",
-  "pink",
-  "brown",
-  "purple",
-  "brown",
-  "white",
-  "black",
-  "orange",
-];
-const N = [
-  "table",
-  "chair",
-  "house",
-  "bbq",
-  "desk",
-  "car",
-  "pony",
-  "cookie",
-  "sandwich",
-  "burger",
-  "pizza",
-  "mouse",
-  "keyboard",
-];
+const A = ["pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean",
+  "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive",
+  "cheap", "expensive", "fancy"];
+const C = ["red", "yellow", "blue", "green", "pink", "brown", "purple", "brown", "white", "black", "orange"];
+const N = ["table", "chair", "house", "bbq", "desk", "car", "pony", "cookie", "sandwich", "burger", "pizza", "mouse",
+  "keyboard"];
 
 let nextId = 1;
 
@@ -67,9 +19,7 @@ const buildData = (count) => {
   for (let i = 0; i < count; i++) {
     data[i] = {
       id: nextId++,
-      label: `${A[random(A.length)]} ${C[random(C.length)]} ${
-        N[random(N.length)]
-      }`,
+      label: `${A[random(A.length)]} ${C[random(C.length)]} ${N[random(N.length)]}`,
     };
   }
 
@@ -78,204 +28,98 @@ const buildData = (count) => {
 
 const initialState = { data: [], selected: 0 };
 
-class Row extends Component {
-  shouldComponentUpdate(nextProps) {
-    const { item, selected } = this.props;
+const listReducer = (state, action) => {
+  const { data, selected } = state;
 
-    return nextProps.item !== item || nextProps.selected !== selected;
+  switch (action.type) {
+    case 'RUN':
+      return { data: buildData(1000), selected: 0 };
+    case 'RUN_LOTS':
+      return { data: buildData(10000), selected: 0 };
+    case 'ADD':
+      return { data: data.concat(buildData(1000)), selected };
+    case 'UPDATE': {
+      const newData = data.slice(0);
+
+      for (let i = 0; i < newData.length; i += 10) {
+        const r = newData[i];
+
+        newData[i] = { id: r.id, label: r.label + " !!!" };
+      }
+
+      return { data: newData, selected };
+    }
+    case 'CLEAR':
+      return { data: [], selected: 0 };
+    case 'SWAP_ROWS':
+      return data.length > 998 ? { data: [data[0], data[998], ...data.slice(2, 998), data[1], data[999]], selected } : state;
+    case 'REMOVE': {
+      const idx = data.findIndex((d) => d.id === action.id);
+
+      return { data: [...data.slice(0, idx), ...data.slice(idx + 1)], selected };
+    }
+    case 'SELECT':
+      return { data, selected: action.id };
+    default:
+      return state;
   }
+};
 
-  constructor(props) {
-    super(props);
+const Row = memo(({ selected, item, dispatch }) => (
+    <tr className={selected ? "danger" : ""}>
+      <td className="col-md-1">{item.id}</td>
+      <td className="col-md-4">
+        <a onClick={() => dispatch({ type: 'SELECT', id: item.id })}>{item.label}</a>
+      </td>
+      <td className="col-md-1">
+        <a onClick={() => dispatch({ type: 'REMOVE', id: item.id })}>
+          <span className="glyphicon glyphicon-remove" aria-hidden="true" />
+        </a>
+      </td>
+      <td className="col-md-6" />
+    </tr>
+), (prevProps, nextProps) => prevProps.selected === nextProps.selected && prevProps.item === nextProps.item)
 
-    this.onSelect = () => {
-      const { item, dispatch } = this.props;
+const Button = ({ id, cb, title }) => (
+  <div className="col-sm-6 smallpad">
+    <button type="button" className="btn btn-primary btn-block" id={id} onClick={cb}>{title}</button>
+  </div>
+);
 
-      dispatch({ type: "SELECT", id: item.id });
-    };
-
-    this.onRemove = () => {
-      const { item, dispatch } = this.props;
-
-      dispatch({ type: "REMOVE", id: item.id });
-    };
-  }
-
-  render() {
-    const { selected, item } = this.props;
-
-    return (
-      <tr className={selected ? "danger" : ""}>
-        <td className="col-md-1">{item.id}</td>
-        <td className="col-md-4">
-          <a onClick={this.onSelect}>{item.label}</a>
-        </td>
-        <td className="col-md-1">
-          <a onClick={this.onRemove}>
-            <span className="glyphicon glyphicon-remove" aria-hidden="true" />
-          </a>
-        </td>
-        <td className="col-md-6" />
-      </tr>
-    );
-  }
-}
-
-class Button extends Component {
-  render() {
-    const { id, cb, title } = this.props;
-
-    return (
-      <div className="col-sm-6 smallpad">
-        <button
-          type="button"
-          className="btn btn-primary btn-block"
-          id={id}
-          onClick={cb}
-        >
-          {title}
-        </button>
-      </div>
-    );
-  }
-}
-
-class Jumbotron extends Component {
-  shouldComponentUpdate() {
-    return false;
-  }
-
-  render() {
-    const { dispatch } = this.props;
-
-    return (
-      <div className="jumbotron">
-        <div className="row">
-          <div className="col-md-6">
-            <h1>React Lab keyed</h1>
-          </div>
-          <div className="col-md-6">
-            <div className="row">
-              <Button
-                id="run"
-                title="Create 1,000 rows"
-                cb={() => dispatch({ type: "RUN" })}
-              />
-              <Button
-                id="runlots"
-                title="Create 10,000 rows"
-                cb={() => dispatch({ type: "RUN_LOTS" })}
-              />
-              <Button
-                id="add"
-                title="Append 1,000 rows"
-                cb={() => dispatch({ type: "ADD" })}
-              />
-              <Button
-                id="update"
-                title="Update every 10th row"
-                cb={() => dispatch({ type: "UPDATE" })}
-              />
-              <Button
-                id="clear"
-                title="Clear"
-                cb={() => dispatch({ type: "CLEAR" })}
-              />
-              <Button
-                id="swaprows"
-                title="Swap Rows"
-                cb={() => dispatch({ type: "SWAP_ROWS" })}
-              />
-            </div>
+const Jumbotron = memo(({ dispatch }) => (
+    <div className="jumbotron">
+      <div className="row">
+        <div className="col-md-6">
+          <h1>React Lab Hooks keyed</h1>
+        </div>
+        <div className="col-md-6">
+          <div className="row">
+            <Button id="run" title="Create 1,000 rows" cb={() => dispatch({ type: 'RUN' })} />
+            <Button id="runlots" title="Create 10,000 rows" cb={() => dispatch({ type: 'RUN_LOTS' })} />
+            <Button id="add" title="Append 1,000 rows" cb={() => dispatch({ type: 'ADD' })} />
+            <Button id="update" title="Update every 10th row" cb={() => dispatch({ type: 'UPDATE' })} />
+            <Button id="clear" title="Clear" cb={() => dispatch({ type: 'CLEAR' })} />
+            <Button id="swaprows" title="Swap Rows" cb={() => dispatch({ type: 'SWAP_ROWS' })} />
           </div>
         </div>
       </div>
-    );
-  }
+    </div>
+), () => true);
+
+const Main = () => {
+  const [{ data, selected }, dispatch] = useReducer(listReducer, initialState);
+
+  return (<div className="container">
+    <Jumbotron dispatch={dispatch} />
+    <table className="table table-hover table-striped test-data">
+      <tbody>
+        {data.map(item => (
+          <Row key={item.id} item={item} selected={selected === item.id} dispatch={dispatch} />
+        ))}
+      </tbody>
+    </table>
+    <span className="preloadicon glyphicon glyphicon-remove" aria-hidden="true" />
+  </div>);
 }
 
-class Main extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = initialState;
-
-    this.dispatch = (action) => {
-      const { data } = this.state;
-
-      switch (action.type) {
-        case "RUN":
-          return this.setState({ data: buildData(1000), selected: 0 });
-        case "RUN_LOTS":
-          return this.setState({ data: buildData(10000), selected: 0 });
-        case "ADD":
-          return this.setState({ data: data.concat(buildData(1000)) });
-        case "UPDATE": {
-          const newData = data.slice(0);
-
-          for (let i = 0; i < newData.length; i += 10) {
-            const r = newData[i];
-
-            newData[i] = { id: r.id, label: r.label + " !!!" };
-          }
-
-          return this.setState({ data: newData });
-        }
-        case "CLEAR":
-          return this.setState({ data: [], selected: 0 });
-        case "SWAP_ROWS": {
-          if (data.length > 998) {
-            return this.setState({
-              data: [
-                data[0],
-                data[998],
-                ...data.slice(2, 998),
-                data[1],
-                data[999],
-              ],
-            });
-          }
-
-          return;
-        }
-        case "REMOVE": {
-          const idx = data.findIndex((d) => d.id === action.id);
-
-          return this.setState({
-            data: [...data.slice(0, idx), ...data.slice(idx + 1)],
-          });
-        }
-        case "SELECT":
-          return this.setState({ selected: action.id });
-      }
-    };
-  }
-
-  render() {
-    const { data, selected } = this.state;
-
-    return (
-      <div className="container">
-        <Jumbotron dispatch={this.dispatch} />
-        <table className="table table-hover table-striped test-data">
-          <tbody>
-            {data.map((item) => (
-              <Row
-                key={item.id}
-                item={item}
-                selected={selected === item.id}
-                dispatch={this.dispatch}
-              />
-            ))}
-          </tbody>
-        </table>
-        <span
-          className="preloadicon glyphicon glyphicon-remove"
-          aria-hidden="true"
-        />
-      </div>
-    );
-  }
-}
-
-createRoot(document.getElementById("main")).render(<Main />);
+createRoot(document.getElementById("main")).render(<Main/>);
