@@ -1,5 +1,5 @@
 import { memo, useCallback } from "react"
-import { render } from "react-dom"
+import { createRoot } from 'react-dom/client';
 import create from "zustand"
 import shallow from "zustand/shallow"
 
@@ -107,9 +107,8 @@ const buildData = (count) => {
 	for (let i = 0; i < count; i++) {
 		data[i] = {
 			id: nextId++,
-			label: `${A[random(A.length)]} ${C[random(C.length)]} ${
-				N[random(N.length)]
-			}`,
+			label: `${A[random(A.length)]} ${C[random(C.length)]} ${N[random(N.length)]
+				}`,
 		}
 	}
 
@@ -231,4 +230,6 @@ const Main = () => {
 	)
 }
 
-render(<Main />, document.getElementById("main"))
+const container = document.getElementById('main');
+const root = createRoot(container);
+root.render(<Main />);
