@@ -1,11 +1,11 @@
 import buble from 'rollup-plugin-buble';
 import nodeResolve from 'rollup-plugin-node-resolve';
-import {uglify} from 'rollup-plugin-uglify';
+import {terser} from 'rollup-plugin-terser';
 
 export default {
   plugins: [
     buble(),
     nodeResolve(),
-    uglify()
+    terser()
   ]
 };
