@@ -3,7 +3,6 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use rand::prelude::*;
 use sycamore::prelude::*;
 use wasm_bindgen::prelude::*;
-use web_sys::window;
 
 static ADJECTIVES: &[&str] = &[
     "pretty",
@@ -43,17 +42,18 @@ static NOUNS: &[&str] = &[
     "pizza", "mouse", "keyboard",
 ];
 
-struct ButtonProps {
+#[derive(Prop)]
+struct ButtonProps<'a> {
     id: &'static str,
     text: &'static str,
-    callback: Box<dyn Fn()>,
+    callback: Box<dyn Fn() + 'a>,
 }
 
-#[component(Button<G>)]
-fn button(props: ButtonProps) -> Template<G> {
+#[component]
+fn Button<'a, G: Html>(cx: Scope<'a>, props: ButtonProps<'a>) -> View<G> {
     let ButtonProps { id, text, callback } = props;
 
-    template! {
+    view! { cx,
         div(class="col-sm-6 smallpad") {
             button(id=id, class="btn btn-primary btn-block", type="button", on:click=move |_| callback()) {
                 (text)
@@ -65,7 +65,7 @@ fn button(props: ButtonProps) -> Template<G> {
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 struct RowData {
     id: usize,
-    label: Signal<String>,
+    label: RcSignal<String>,
 }
 
 static ID_COUNTER: AtomicUsize = AtomicUsize::new(1);
@@ -90,7 +90,7 @@ fn build_data(count: usize) -> Vec<RowData> {
 
         data.push(RowData {
             id: ID_COUNTER.load(Ordering::Relaxed),
-            label: Signal::new(label),
+            label: create_rc_signal(label),
         });
 
         ID_COUNTER.store(ID_COUNTER.load(Ordering::Relaxed) + 1, Ordering::Relaxed);
@@ -99,91 +99,97 @@ fn build_data(count: usize) -> Vec<RowData> {
     data
 }
 
-#[component(App<G>)]
-fn app() -> Template<G> {
-    let data = Signal::new(Vec::<RowData>::new());
-    let selected = Signal::new(None::<usize>);
+#[component]
+fn App<G: Html>(cx: Scope) -> View<G> {
+    let data = create_signal(cx, Vec::<RowData>::new());
+    let selected = create_signal(cx, None::<usize>);
 
-    let remove = cloned!((data) => move |id| {
-        data.set(data.get().iter().filter(|row| row.id != id).cloned().collect());
-    });
+    let remove = |id| {
+        data.set(
+            data.get()
+                .iter()
+                .filter(|row| row.id != id)
+                .cloned()
+                .collect(),
+        );
+    };
 
-    let run = cloned!((data, selected) => move || {
+    let run = || {
         data.set(build_data(1000));
         selected.set(None);
-    });
+    };
 
-    let runlots = cloned!((data, selected) => move || {
+    let runlots = || {
         data.set(build_data(10000));
         selected.set(None);
-    });
+    };
 
-    let add = cloned!((data) => move || {
+    let add = || {
         data.set(data.get().iter().cloned().chain(build_data(1000)).collect());
-    });
+    };
 
-    let update = cloned!((data) => move || {
+    let update = || {
         let mut tmp = (*data.get()).clone();
         for row in tmp.iter_mut().step_by(10) {
             row.label.set(format!("{} !!!", row.label.get()));
         }
         data.set(tmp);
-    });
+    };
 
-    let clear = cloned!((data, selected) => move || {
+    let clear = || {
         data.set(Vec::new());
         selected.set(None);
-    });
+    };
 
-    let swaprows = cloned!((data) => move || {
+    let swaprows = || {
         let mut d = (*data.get()).clone();
         if d.len() > 998 {
             d.swap(1, 998);
         }
         data.set(d);
-    });
+    };
 
-    template! {
+    view! { cx,
         div(class="container") {
             div(class="jumbotron") {
                 div(class="row") {
                     div(class="col-md-6") { h1 { "Sycamore Keyed" } }
                     div(class="col-md-6") {
                         div(class="row") {
-                            Button(ButtonProps { id: "run", text: "Create 1,000 rows", callback: Box::new(run) })
-                            Button(ButtonProps { id: "runlots", text: "Create 10,000 rows", callback: Box::new(runlots) })
-                            Button(ButtonProps { id: "add", text: "Append 1,000 rows", callback: Box::new(add) })
-                            Button(ButtonProps { id: "update", text: "Update every 10th row", callback: Box::new(update) })
-                            Button(ButtonProps { id: "clear", text: "Clear", callback: Box::new(clear) })
-                            Button(ButtonProps { id: "swaprows", text: "Swap Rows", callback: Box::new(swaprows) })
+                            Button(id="run", text="Create 1,000 rows", callback=Box::new(run))
+                            Button(id="runlots", text="Create 10,000 rows", callback=Box::new(runlots))
+                            Button(id="add", text="Append 1,000 rows", callback=Box::new(add))
+                            Button(id="update", text="Update every 10th row", callback=Box::new(update))
+                            Button(id="clear", text="Clear", callback=Box::new(clear))
+                            Button(id="swaprows", text="Swap Rows", callback=Box::new(swaprows))
                         }
                     }
                 }
             }
             table(class="table table-hover table-striped test-data") {
                 tbody {
-                    Keyed(KeyedProps {
-                        iterable: data.handle(),
-                        template: move |row| {
-                            let row_id = row.id;
-                            let is_selected = create_selector(cloned!((selected) => move || *selected.get() == Some(row_id)));
-                            cloned!((selected, remove) => template! {
-                                tr(class=if *is_selected.get() { "danger" } else { "" }) {
-                                    td(class="col-md-1") { (row_id) }
+                    Keyed(
+                        iterable=data,
+                        view=move |cx, row| {
+                            let is_selected = create_selector(cx, move || *selected.get() == Some(row.id));
+                            let handle_click = move |_| selected.set(Some(row.id));
+                            view! { cx,
+                                tr(class=is_selected.get().then(|| "danger").unwrap_or("")) {
+                                    td(class="col-md-1") { (row.id) }
                                     td(class="col-md-4") {
-                                        a(on:click=move |_| selected.set(Some(row_id))) { (row.label.get()) }
+                                        a(on:click=handle_click) { (row.label.get()) }
                                     }
                                     td(class="col-md-1") {
-                                        a(on:click=move |_| remove(row_id)) {
+                                        a(on:click=move |_| remove(row.id)) {
                                             span(class="glyphicon glyphicon-remove", aria-hidden="true")
                                         }
                                     }
                                     td(class="col-md-6")
                                 }
-                            })
+                            }
                         },
-                        key: |row| row.id
-                    })
+                        key=|row| row.id
+                    )
                 }
             }
         }
@@ -192,7 +198,7 @@ fn app() -> Template<G> {
 
 #[wasm_bindgen(start)]
 pub fn start() {
-    let document = window().unwrap().document().unwrap();
+    let document = web_sys::window().unwrap().document().unwrap();
     let mount_el = document.query_selector("#main").unwrap().unwrap();
-    sycamore::render_to(|| template! { App() }, &mount_el);
+    sycamore::render_to(|cx| view! { cx, App {} }, &mount_el);
 }
