@@ -1,1 +1,1 @@
-var b=Symbol("value"),$=Symbol("name"),u=new WeakMap,g=new WeakMap,p=new Map,v=[],S=m({});function r(e,t,...s){if(j(e))return e({...t,children:s});if(C(e))return r("",void 0,e.children);let n=e===""?document.createDocumentFragment():document.createElement(e);for(let[o,d]of Object.entries(t??{}))o in n?n[o]=d:n.setAttribute(o,String(d));n.append(...s.flat(Infinity));let[l,i,c]=v.pop()||[];if(l){let o=s.flatMap(G);try{if(o.includes(c)||(!o.includes(c)&&V(n)?n.textContent.includes(c):n.outerHTML.includes(c))){let d=V(n)?n.firstChild:n;u.has(l)?u.get(l).push(d):u.set(l,[d]),g.set(d,[i,c])}}catch{return n}}return n}function N(e,t){t?t.replaceWith(e):document.body.insertBefore(e,null)}var h=[],x=!1;function m(e){Reflect.set(n,b,e);let t=new Proxy(n,{set(l,i,c,o){let d=Reflect.get(l,i,o);if(d===c)return!0;let a=Reflect.set(l,i,c,o);return d!=null&&a&&!x&&z(F,c,o),p.has(o)&&p.get(o).forEach(O.bind(O,c,d)),a},get(l,i,c){let o=Reflect.get(l,b),d=C(o),a;d&&(a=Reflect.get(o,i,c)||o);let y=[c,i,d?a:o];return v.push(y),queueMicrotask(H.bind(H,y)),d?a:P.bind(P,o)}}),s=K();return Reflect.set(t,$,s),S&&!Reflect.has(S,s)&&(x=!0,Reflect.set(S,s,t),x=!1),t;function n(l){if(t!==null){if(l===null){Reflect.deleteProperty(t,b);let i;p.has(t)&&p.delete(t),u.has(t)&&(i=u.get(t),u.delete(t));for(let c=0;i&&c<i.length;c++){let o=i[c];g.has(o)&&g.delete(o)}t=l;return}else if(j(l)){Reflect.has(l,b)?(h.push(l),h.length===2&&(X(h[0],h[1]),h=[])):n(l(Reflect.get(t,b)));return}Reflect.set(t,b,l)}}}function R(e,t){p.has(e)?p.get(e).push(t):p.set(e,[t])}function O(e,t,s){s(e,t)}function E(e,t,s){let n=document.body.querySelector(e),l=Reflect.get(t,b).map(s);n.append(...l),R(t,T.bind(T,n,s))}function T(e,t,s,n){let l=new Set([...n,...s]).size===n?.length+s?.length;if(!s?.length||n?.length===s?.length||l){e.textContent="";for(let i=0;i<n.length;i++)n[i](null)}if(n?.length&&s?.length>n?.length){let i=n.length,o=(l?s:s.slice(i)).map(A.bind(A,t,i));e.append(...o)}else if(s?.length&&s?.length<n.length)n.reverse().forEach(q.bind(q,e,s,n.length-1)),s.forEach(I.bind(I,e,t,n));else if(n?.length===0||n?.length===s?.length){let i=s.map(t);e.append(...i)}}function A(e,t,s,n){return e(s,n+t)}function I(e,t,s,n,l){s.includes(n)||e.appendChild(t(n,l))}function q(e,t,s,n,l){t.includes(n)||(n(null),e.childNodes[s-l]?.remove())}function F(e,t){if(u.has(t)){let s=u.get(t);for(let n=0;n<s.length;n++){let l=s[n],[i,c]=g.get(l),o=C(e)?Reflect.get(e,i):e,d=!1;!Q(l)&&l.outerHTML.includes(c)&&(d=l.getAttributeNames().some(L.bind(L,l,c,o))),d||(l.textContent=l.textContent.replace(c,o)),g.set(l,[i,o])}}}function L(e,t,s,n){if(e.getAttribute(n)===t)return e.setAttribute(n,s),!0}function z(e,...t){navigator.scheduling?navigator.scheduling.isInputPending()?setTimeout(z,0,e,...t):e(...t):(requestIdleCallback||setTimeout)(e.bind(e,...t))}function k(e){return Reflect.get(e,b)}function X(e,t){let s=new Set,n=u.get(e),l=u.get(t);for(let i=0;i<n.length;i++){let c=n[i].closest("[data-bind]")||n[i],o=l[i].closest("[data-bind]")||l[i];if(s.has(c)||s.has(o))continue;s.add(c),s.add(o);let d=o.nextSibling,a=o.previousSibling,y=o.parentNode;c.before(o),d?d.before(c):a?a.after(c):y.appendChild(c)}}function P(e){return e}function G(e){return String(e)}function U(e,t){return e===t}function K(){return Math.random().toString(32).slice(2)}function C(e){return e!=null&&typeof e=="object"}function j(e){return typeof e=="function"}function H(e){let t=v.findIndex(U.bind(U,e));v.splice(t,1)}function Q(e){return e.splitText!==void 0}function V(e){return e.nodeName!=="svg"&&"getElementById"in e}var B=["pretty","large","big","small","tall","short","long","handsome","plain","quaint","clean","elegant","easy","angry","crazy","helpful","mushy","odd","unsightly","adorable","important","inexpensive","cheap","expensive","fancy"],J=["red","yellow","blue","green","pink","brown","purple","brown","white","black","orange"],W=["table","chair","house","bbq","desk","car","pony","cookie","sandwich","burger","pizza","mouse","keyboard"],Y=B.length,Z=J.length,D=W.length,ee=1;function M(e){let t=new Array(e);for(let s=0;s<e;s++)t[s]=m({id:ee++,label:B[w(Y)]+" "+J[w(Z)]+" "+W[w(D)]});return t}function w(e){return Math.random()*e|0}var te=document.querySelector.bind(document);N(r("div",{class:"row"},r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"run",onclick:se},"Create 1,000 rows")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"runlots",onclick:le},"Create 10,000 rows")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"add",onclick:ne},"Append 1,000 rows")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"update",onclick:ie},"Update every 10th row")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"clear",onclick:oe},"Clear")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"swaprows",onclick:ce},"Swap Rows"))),te(".col-md-6 > .row"));var f=m([]),_=m(-1);E("tbody",f,e=>{let t=m(""),s=r("tr",{class:t,"data-bind":!0},r("td",{class:"col-md-1"},e.id),r("td",{class:"col-md-4"},r("a",{onclick:()=>_(e.id)},e.label)),r("td",{class:"col-md-1"},r("a",{onclick:()=>re(e.id)},r("span",{class:"glyphicon glyphicon-remove","aria-hidden":"true"}))),r("td",{class:"col-md-6"}));return R(_,n=>{n===e.id?t("danger"):t("")}),s});function ne(){f(e=>[...e,...M(1e3)])}function se(){f(M(1e3))}function le(){f(M(1e4))}function oe(){f([])}function ie(){let e=k(f),t=0;for(;t<e.length;)e[t](s=>({...s,label:s.label+=" !!!"})),t+=10}function ce(){let e=k(f);if(e.length>998){let t=e[1];e[1](e[998]),e[998](t)}}function re(e){let t=[...k(f)],s=t.findIndex(n=>n?.id===e);t.splice(s,1),f(t)}
+"use strict";var b=Symbol("value"),$=Symbol("name"),u=new WeakMap,h=new WeakMap,p=new Map,v=[],S=m({});function r(e,t,...s){if(V(e))return e({...t,children:s});if(C(e))return r("",void 0,e.children);let n=e===""?document.createDocumentFragment():document.createElement(e);for(let[o,d]of Object.entries(t??{}))o in n?n[o]=d:n.setAttribute(o,String(d));n.append(...s.flat(1/0));let[l,i,c]=v.pop()||[];if(l){let o=s.flatMap(G);try{if(o.includes(c)||(!o.includes(c)&&P(n)?n.textContent.includes(c):n.outerHTML.includes(c))){let d=P(n)?n.firstChild:n;u.has(l)?u.get(l).push(d):u.set(l,[d]),h.set(d,[i,c])}}catch{return n}}return n}function U(e,t){t?t.replaceWith(e):document.body.insertBefore(e,null)}var g=[],x=!1;function m(e){Reflect.set(n,b,e);let t=new Proxy(n,{set(l,i,c,o){let d=Reflect.get(l,i,o);if(d===c)return!0;let a=Reflect.set(l,i,c,o);return d!=null&&a&&!x&&H(F,c,o),p.has(o)&&p.get(o).forEach(N.bind(N,c,d)),a},get(l,i,c){let o=Reflect.get(l,b),d=C(o),a;d&&(a=Reflect.get(o,i,c)||o);let y=[c,i,d?a:o];return v.push(y),queueMicrotask(z.bind(z,y)),d?a:q.bind(q,o)}}),s=K();return Reflect.set(t,$,s),S&&!Reflect.has(S,s)&&(x=!0,Reflect.set(S,s,t),x=!1),t;function n(l){if(t!==null){if(l===null){Reflect.deleteProperty(t,b);let i;p.has(t)&&p.delete(t),u.has(t)&&(i=u.get(t),u.delete(t));for(let c=0;i&&c<i.length;c++){let o=i[c];h.has(o)&&h.delete(o)}t=l;return}else if(V(l)){Reflect.has(l,b)?(g.push(l),g.length===2&&(X(g[0],g[1]),g=[])):n(l(Reflect.get(t,b)));return}Reflect.set(t,b,l)}}}function R(e,t){p.has(e)?p.get(e).push(t):p.set(e,[t])}function N(e,t,s){s(e,t)}function j(e,t,s){let n=document.body.querySelector(e),l=Reflect.get(t,b).map(s);n.append(...l),R(t,O.bind(O,n,s))}function O(e,t,s,n){let l=new Set([...n,...s]).size===n?.length+s?.length;if(!s?.length||n?.length===s?.length||l){e.textContent="";for(let i=0;i<n.length;i++)n[i](null)}if(n?.length&&s?.length>n?.length){let i=n.length,o=(l?s:s.slice(i)).map(E.bind(E,t,i));e.append(...o)}else if(s?.length&&s?.length<n.length)n.reverse().forEach(A.bind(A,e,s,n.length-1)),s.forEach(T.bind(T,e,t,n));else if(n?.length===0||n?.length===s?.length){let i=s.map(t);e.append(...i)}}function E(e,t,s,n){return e(s,n+t)}function T(e,t,s,n,l){s.includes(n)||e.appendChild(t(n,l))}function A(e,t,s,n,l){t.includes(n)||(n(null),e.childNodes[s-l]?.remove())}function F(e,t){if(u.has(t)){let s=u.get(t);for(let n=0;n<s.length;n++){let l=s[n],[i,c]=h.get(l),o=C(e)?Reflect.get(e,i):e,d=!1;!Q(l)&&l.outerHTML.includes(c)&&(d=l.getAttributeNames().some(I.bind(I,l,c,o))),d||(l.textContent=l.textContent.replace(c,o)),h.set(l,[i,o])}}}function I(e,t,s,n){if(e.getAttribute(n)===t)return e.setAttribute(n,s),!0}function H(e,...t){navigator.scheduling?navigator.scheduling.isInputPending()?setTimeout(H,0,e,...t):e(...t):(requestIdleCallback||setTimeout)(e.bind(e,...t))}function k(e){return Reflect.get(e,b)}function X(e,t){let s=new Set,n=u.get(e),l=u.get(t);for(let i=0;i<n.length;i++){let c=n[i].closest("[data-bind]")||n[i],o=l[i].closest("[data-bind]")||l[i];if(s.has(c)||s.has(o))continue;s.add(c),s.add(o);let d=o.nextSibling,a=o.previousSibling,y=o.parentNode;c.before(o),d?d.before(c):a?a.after(c):y.appendChild(c)}}function q(e){return e}function G(e){return String(e)}function L(e,t){return e===t}function K(){return Math.random().toString(32).slice(2)}function C(e){return e!=null&&typeof e=="object"}function V(e){return typeof e=="function"}function z(e){let t=v.findIndex(L.bind(L,e));v.splice(t,1)}function Q(e){return e.splitText!==void 0}function P(e){return e.nodeName!=="svg"&&"getElementById"in e}var J=["pretty","large","big","small","tall","short","long","handsome","plain","quaint","clean","elegant","easy","angry","crazy","helpful","mushy","odd","unsightly","adorable","important","inexpensive","cheap","expensive","fancy"],W=["red","yellow","blue","green","pink","brown","purple","brown","white","black","orange"],_=["table","chair","house","bbq","desk","car","pony","cookie","sandwich","burger","pizza","mouse","keyboard"],Y=J.length,Z=W.length,D=_.length,ee=1;function w(e){let t=new Array(e);for(let s=0;s<e;s++)t[s]=m({id:ee++,label:J[M(Y)]+" "+W[M(Z)]+" "+_[M(D)]});return t}function M(e){return Math.random()*e|0}var te=document.querySelector.bind(document);U(r("div",{class:"row"},r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"run",onclick:se},"Create 1,000 rows")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"runlots",onclick:le},"Create 10,000 rows")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"add",onclick:ne},"Append 1,000 rows")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"update",onclick:ie},"Update every 10th row")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"clear",onclick:oe},"Clear")),r("div",{class:"col-sm-6 smallpad"},r("button",{type:"button",class:"btn btn-primary btn-block",id:"swaprows",onclick:ce},"Swap Rows"))),te(".col-md-6 > .row"));var f=m([]),B=m(-1);j("tbody",f,e=>{let t=m(""),s=r("tr",{class:t,"data-bind":!0},r("td",{class:"col-md-1"},e.id),r("td",{class:"col-md-4"},r("a",{onclick:()=>B(e.id)},e.label)),r("td",{class:"col-md-1"},r("a",{onclick:()=>re(e.id)},r("span",{class:"glyphicon glyphicon-remove","aria-hidden":"true"}))),r("td",{class:"col-md-6"}));return R(B,n=>{n===e.id?t("danger"):t("")}),s});function ne(){f(e=>[...e,...w(1e3)])}function se(){f(w(1e3))}function le(){f(w(1e4))}function oe(){f([])}function ie(){let e=k(f),t=0;for(;t<e.length;)e[t](s=>({...s,label:s.label+=" !!!"})),t+=10}function ce(){let e=k(f);if(e.length>998){let t=e[1];e[1](e[998]),e[998](t)}}function re(e){let t=[...k(f)],s=t.findIndex(n=>n?.id===e);t.splice(s,1),f(t)}
