@@ -1,32 +1,15 @@
-// import { BenchmarkInfo, BenchmarkType } from "./benchmarksCommon";
-// import { config, FrameworkData } from "./common";
-// import * as benchmarksCommon from "./benchmarksCommon";
-// import {DurationMeasurementMode} from "./benchmarksCommon";
-
-import { BenchmarkImpl, BenchmarkType, BENCHMARK_30, DurationMeasurementMode, StartupBenchmark, StartupMainBenchmark } from "./benchmarksCommon";
-
-
-// export interface LighthouseData {
-//   TimeToConsistentlyInteractive: number;
-//   ScriptBootUpTtime: number;
-//   MainThreadWorkCost: number;
-//   TotalKiloByteWeight: number;
-//   [propName: string]: number;
-// }
-
-// export interface StartupBenchmarkResult extends BenchmarkInfo {
-//   property: keyof LighthouseData;
-// }
+import * as benchmarksCommon from "./benchmarksCommon";
+import { BenchmarkImpl, BenchmarkType, StartupBenchmarkInfo } from "./benchmarksCommon";
 
 export interface StartupBenchmarkResult extends BenchmarkImpl {
-  benchmark: StartupBenchmark;
+  benchmark: StartupBenchmarkInfo;
   result: number;
 }
 
 let id = (x:number) => x;
 let toKb = (x:number) => x/1024;
 
-export const benchStartupConsistentlyInteractive: StartupBenchmark = {
+export const benchStartupConsistentlyInteractive: StartupBenchmarkInfo = {
   id: "31_startup-ci",
   label: "consistently interactive",
   description: "a pessimistic TTI - when the CPU and network are both definitely very idle. (no more CPU tasks over 50ms)",
@@ -35,7 +18,7 @@ export const benchStartupConsistentlyInteractive: StartupBenchmark = {
   type: BenchmarkType.STARTUP,
 };
 
-export const benchStartupBootup: StartupBenchmark = {
+export const benchStartupBootup: StartupBenchmarkInfo = {
   id: "32_startup-bt",
   label: "script bootup time",
   description: "the total ms required to parse/compile/evaluate all the page's scripts",
@@ -44,7 +27,7 @@ export const benchStartupBootup: StartupBenchmark = {
   type: BenchmarkType.STARTUP,
 };
 
-export const benchStartupMainThreadWorkCost: StartupBenchmark = {
+export const benchStartupMainThreadWorkCost: StartupBenchmarkInfo = {
   id: "33_startup-mainthreadcost",
   label: "main thread work cost",
   description: "total amount of time spent doing work on the main thread. includes style/layout/etc.",
@@ -53,7 +36,7 @@ export const benchStartupMainThreadWorkCost: StartupBenchmark = {
   type: BenchmarkType.STARTUP,
 };
 
-export const benchStartupTotalBytes: StartupBenchmark = {
+export const benchStartupTotalBytes: StartupBenchmarkInfo = {
   id: "34_startup-totalbytes",
   label: "total kilobyte weight",
   description: "network transfer cost (post-compression) of all the resources loaded into the page.",
@@ -62,18 +45,14 @@ export const benchStartupTotalBytes: StartupBenchmark = {
   type: BenchmarkType.STARTUP,
 };
 
+export const subbenchmarks = [benchStartupConsistentlyInteractive, benchStartupBootup, benchStartupMainThreadWorkCost, benchStartupTotalBytes];
+
 export class BenchmarkLighthouse implements BenchmarkImpl {
   type = BenchmarkType.STARTUP_MAIN;
-
-  subbenchmarks = [benchStartupConsistentlyInteractive, benchStartupBootup, benchStartupMainThreadWorkCost, benchStartupTotalBytes];
-
-  constructor(public benchmarkInfo: StartupMainBenchmark) {
-  }
+  benchmarkInfo = benchmarksCommon.startupBenchmarkInfos[benchmarksCommon.BENCHMARK_30];
+  subbenchmarks = subbenchmarks;
 }
 
-export const benchLighthouse = new BenchmarkLighthouse({
-  id: BENCHMARK_30,
-  type: BenchmarkType.STARTUP_MAIN,
-  label: '',
-  description: '',
-});
+export const benchLighthouse = new BenchmarkLighthouse();
+
+export const benchmarks = [benchLighthouse];
