@@ -1,66 +1,47 @@
+import { WebDriver } from "selenium-webdriver";
+import * as benchmarksCommon from "./benchmarksCommon";
 import { BenchmarkType } from "./benchmarksCommon";
+import { config, FrameworkData } from "./common";
 import {
-  testTextContains,
-  testTextNotContained,
-  testClassContains,
-  testElementLocatedByXpath,
-  testElementNotLocatedByXPath,
-  testElementLocatedById,
   clickElementById,
   clickElementByXPath,
-  getTextByXPath,
+  getTextByXPath, testClassContains, testElementLocatedById, testElementLocatedByXpath,
+  testElementNotLocatedByXPath, testTextContains
 } from "./webdriverAccess";
-import { Builder, WebDriver, promise, logging } from "selenium-webdriver";
-import { config, FrameworkData } from "./common";
-import * as benchmarksCommon from "./benchmarksCommon";
-import {slowDownFactor, slowDownNote, DurationMeasurementMode} from "./benchmarksCommon";
 
 
 const SHORT_TIMEOUT = 20 * 1000;
 
 export abstract class CPUBenchmarkWebdriver {
   type = BenchmarkType.CPU;
-  constructor(public benchmarkInfo: benchmarksCommon.CPUBenchmark) {
+  constructor(public benchmarkInfo: benchmarksCommon.CPUBenchmarkInfo) {
   }
   abstract init(driver: WebDriver, framework: FrameworkData): Promise<any>;
   abstract run(driver: WebDriver, framework: FrameworkData): Promise<any>;
-  after(driver: WebDriver, framework: FrameworkData): Promise<any> {
-    return null;
-  }
 }
 
 export const benchRun = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_01,
-      label: "create rows",
-      description: "creating 1,000 rows" + slowDownNote(benchmarksCommon.BENCHMARK_01),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_01),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.FIRST_PAINT_AFTER_LAYOUT
-    });
-  }
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_01]);
+  } 
   async init(driver: WebDriver) {
-    await testElementLocatedById(driver, "add", SHORT_TIMEOUT, true);
+    await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
+    for (let i = 0; i < config.WARMUP_COUNT; i++) {
+      await clickElementById(driver, "run", true);
+      await testTextContains(driver, "//tbody/tr[1]/td[1]", (i * 1000 + 1).toFixed(), config.TIMEOUT, false);
+      await clickElementById(driver, "clear", true);
+      await testElementNotLocatedByXPath(driver, "//tbody/tr[1]", config.TIMEOUT, false);
+    }
   }
   async run(driver: WebDriver) {
-    await clickElementById(driver, "add", true);
-    await testElementLocatedByXpath(driver, "//tbody/tr[1000]/td[2]/a", config.TIMEOUT, false);
+    await clickElementById(driver, "run", true);
+    await testTextContains(driver, "//tbody/tr[1]/td[1]", (config.WARMUP_COUNT * 1000 + 1).toFixed(), config.TIMEOUT, false);
   }
 })();
 
 export const benchReplaceAll = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_02,
-      label: "replace all rows",
-      description: "updating all 1,000 rows (" + config.WARMUP_COUNT + " warmup runs)." + slowDownNote(benchmarksCommon.BENCHMARK_02),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_02),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.FIRST_PAINT_AFTER_LAYOUT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_02]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
@@ -77,15 +58,7 @@ export const benchReplaceAll = new (class extends CPUBenchmarkWebdriver {
 
 export const benchUpdate = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_03,
-      label: "partial update",
-      description: "updating every 10th row for 1,000 rows (3 warmup runs)." + slowDownNote(benchmarksCommon.BENCHMARK_03),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_03),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.FIRST_PAINT_AFTER_LAYOUT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_03]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
@@ -104,15 +77,7 @@ export const benchUpdate = new (class extends CPUBenchmarkWebdriver {
 
 export const benchSelect = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_04,
-      label: "select row",
-      description: "highlighting a selected row. (no warmup runs)." + slowDownNote(benchmarksCommon.BENCHMARK_04),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_04),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.LAST_PAINT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_04]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
@@ -127,15 +92,7 @@ export const benchSelect = new (class extends CPUBenchmarkWebdriver {
 
 export const benchSwapRows = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_05,
-      label: "swap rows",
-      description: "swap 2 rows for table with 1,000 rows. (" + config.WARMUP_COUNT + " warmup runs)." + slowDownNote(benchmarksCommon.BENCHMARK_05),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_05),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.LAST_PAINT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_05]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
@@ -156,15 +113,7 @@ export const benchSwapRows = new (class extends CPUBenchmarkWebdriver {
 
 export const benchRemove = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_06,
-      label: "remove row",
-      description: "removing one row. (" + config.WARMUP_COUNT + " warmup runs)." + slowDownNote(benchmarksCommon.BENCHMARK_06),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_06),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.LAST_PAINT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_06]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
@@ -197,18 +146,16 @@ export const benchRemove = new (class extends CPUBenchmarkWebdriver {
 
 export const benchRunBig = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_07,
-      label: "create many rows" + slowDownNote(benchmarksCommon.BENCHMARK_07),
-      description: "creating 10,000 rows",
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_07),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.FIRST_PAINT_AFTER_LAYOUT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_07]);
   }
   async init(driver: WebDriver) {
-    await testElementLocatedById(driver, "runlots", SHORT_TIMEOUT, true);
+    await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
+    for (let i = 0; i < config.WARMUP_COUNT; i++) {
+      await clickElementById(driver, "run", true);
+      await testTextContains(driver, "//tbody/tr[1]/td[1]", (i * 1000 + 1).toFixed(), config.TIMEOUT, false);
+      await clickElementById(driver, "clear", true);
+      await testElementNotLocatedByXPath(driver, "//tbody/tr[1]", config.TIMEOUT, false);
+    }
   }
   async run(driver: WebDriver) {
     await clickElementById(driver, "runlots", true);
@@ -218,15 +165,7 @@ export const benchRunBig = new (class extends CPUBenchmarkWebdriver {
 
 export const benchAppendToManyRows = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_08,
-      label: "append rows to table",
-      description: "appending 1,000 to a table of 1,000 rows." + slowDownNote(benchmarksCommon.BENCHMARK_08),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_08),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.FIRST_PAINT_AFTER_LAYOUT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_08]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
@@ -241,18 +180,16 @@ export const benchAppendToManyRows = new (class extends CPUBenchmarkWebdriver {
 
 export const benchClear = new (class extends CPUBenchmarkWebdriver {
   constructor() {
-    super({
-      id: benchmarksCommon.BENCHMARK_09,
-      label: "clear rows",
-      description: "clearing a table with 1,000 rows." + slowDownNote(benchmarksCommon.BENCHMARK_09),
-      type: BenchmarkType.CPU,
-      throttleCPU: slowDownFactor(benchmarksCommon.BENCHMARK_09),
-      allowBatching: true,
-      durationMeasurementMode: DurationMeasurementMode.FIRST_PAINT_AFTER_LAYOUT
-    });
+    super(benchmarksCommon.cpuBenchmarkInfos[benchmarksCommon.BENCHMARK_09]);
   }
   async init(driver: WebDriver) {
     await testElementLocatedById(driver, "run", SHORT_TIMEOUT, true);
+    for (let i = 0; i < config.WARMUP_COUNT; i++) {
+      await clickElementById(driver, "run", true);
+      await testElementLocatedByXpath(driver, "//tbody/tr[1000]/td[2]/a", config.TIMEOUT, false);
+      await clickElementById(driver, "clear", true);
+      await testElementNotLocatedByXPath(driver, "//tbody/tr[1]", config.TIMEOUT, false);
+    }
     await clickElementById(driver, "run", true);
     await testElementLocatedByXpath(driver, "//tbody/tr[1000]/td[2]/a", config.TIMEOUT, false);
   }
@@ -261,3 +198,15 @@ export const benchClear = new (class extends CPUBenchmarkWebdriver {
     await testElementNotLocatedByXPath(driver, "//tbody/tr[1]", config.TIMEOUT, false);
   }
 })();
+
+export const benchmarks = [
+  benchRun, 
+  benchReplaceAll,
+  benchUpdate, 
+  benchSelect, 
+  benchSwapRows, 
+  benchRemove, 
+  benchRunBig, 
+  benchAppendToManyRows, 
+  benchClear, 
+];
