@@ -1,7 +1,9 @@
 import * as fs from "fs";
 import * as path from "path";
-import axios from "axios";
+import * as axios from "axios";
 import { StartupBenchmarkResult } from "./benchmarksLighthouse";
+import { string } from "yargs";
+
 
 export interface JSONResult {
   framework: string;
@@ -20,8 +22,8 @@ export interface JSONResult {
 export type TBenchmarkStatus = "OK" | "TEST_FAILED" | "TECHNICAL_ERROR";
 
 export interface ErrorAndWarning {
-  error: String;
-  warnings: String[];
+  error: string;
+  warnings: string[];
   result?: number[] | StartupBenchmarkResult[];
 }
 
@@ -33,11 +35,19 @@ export interface BenchmarkDriverOptions {
 }
 
 export interface BenchmarkOptions extends BenchmarkDriverOptions {
+  HOST: string;
   port: string;
   batchSize: number;
   numIterationsForCPUBenchmarks: number;
   numIterationsForMemBenchmarks: number;
   numIterationsForStartupBenchmark: number;
+}
+
+export enum BENCHMARK_RUNNER { 
+  PUPPETEER = "puppeteer", 
+  PLAYWRIGHT = "playwright", 
+  WEBDRIVER_CDP = "webdrivercdp", 
+  WEBDRIVER = "webdriver" 
 }
 
 export let config = {
@@ -61,6 +71,8 @@ export let config = {
   RESULTS_DIRECTORY: "results",
   TRACES_DIRECTORY: "traces",
   ALLOW_BATCHING: true,
+  HOST: 'localhost',
+  BENCHMARK_RUNNER: BENCHMARK_RUNNER.PUPPETEER
 };
 export type TConfig = typeof config;
 
@@ -87,101 +99,18 @@ function computeHash(keyedType: KeyedType, directory: string) {
   return keyedType + "/" + directory;
 }
 
-export interface FrameworkId {
-  keyedType: KeyedType;
+export interface FrameworkInformation {
+  type: KeyedType;
   directory: string;
-  issues: number[];
+  issues: string[];
+  customURL?: string;
+  useShadowRoot?: boolean;
+  useRowShadowRoot?: boolean;
+  shadowRootName?: string;
+  buttonsInShadowRoot?: boolean;
+  versions?: {[key: string]: string};
+  frameworkVersionString: string;
 }
-
-abstract class FrameworkVersionInformationValid implements FrameworkId {
-  public url: string;
-  constructor(
-    public keyedType: KeyedType,
-    public directory: string,
-    customURL: string | undefined,
-    public useShadowRoot: boolean,
-    public useRowShadowRoot: boolean,
-    public shadowRootName: string,
-    public buttonsInShadowRoot: boolean,
-    public issues: number[]
-  ) {
-    this.keyedType = keyedType;
-    this.directory = directory;
-    this.url = "frameworks/" + keyedType + "/" + directory + (customURL ? customURL : "");
-  }
-}
-
-export class FrameworkVersionInformationDynamic extends FrameworkVersionInformationValid {
-  constructor(
-    keyedType: KeyedType,
-    directory: string,
-    public packageNames: string[],
-    customURL: string | undefined,
-    useShadowRoot: boolean = false,
-    useRowShadowRoot: boolean = false,
-    public shadowRootName: string,
-    public buttonsInShadowRoot: boolean,
-    issues: number[]
-  ) {
-    super(keyedType, directory, customURL, useShadowRoot, useRowShadowRoot, shadowRootName, buttonsInShadowRoot, issues);
-  }
-}
-
-export class FrameworkVersionInformationStatic extends FrameworkVersionInformationValid {
-  constructor(
-    keyedType: KeyedType,
-    directory: string,
-    public frameworkVersion: string,
-    customURL: string | undefined,
-    useShadowRoot: boolean = false,
-    useRowShadowRoot: boolean = false,
-    public shadowRootName: string,
-    public buttonsInShadowRoot: boolean,
-    issues: number[]
-  ) {
-    super(keyedType, directory, customURL, useShadowRoot, useRowShadowRoot, shadowRootName, buttonsInShadowRoot, issues);
-  }
-  getFrameworkData(): FrameworkData {
-    return {
-      name: this.directory,
-      fullNameWithKeyedAndVersion: this.directory + (this.frameworkVersion ? "-v" + this.frameworkVersion : "") + "-" + this.keyedType,
-      uri: this.url,
-      keyed: this.keyedType === "keyed",
-      useShadowRoot: this.useShadowRoot,
-      useRowShadowRoot: this.useRowShadowRoot,
-      shadowRootName: this.shadowRootName,
-      buttonsInShadowRoot: this.buttonsInShadowRoot,
-      issues: this.issues,
-    };
-  }
-}
-
-export class FrameworkVersionInformationError implements FrameworkId {
-  public issues: [];
-  constructor(public keyedType: KeyedType, public directory: string, public error: string) {}
-}
-
-export type FrameworkVersionInformation =
-  | FrameworkVersionInformationDynamic
-  | FrameworkVersionInformationStatic
-  | FrameworkVersionInformationError;
-
-export class PackageVersionInformationValid {
-  constructor(public packageName: string, public version: string) {}
-}
-
-export class PackageVersionInformationErrorUnknownPackage {
-  constructor(public packageName: string) {}
-}
-
-export class PackageVersionInformationErrorNoPackageJSONLock {
-  constructor() {}
-}
-
-export type PackageVersionInformation =
-  | PackageVersionInformationValid
-  | PackageVersionInformationErrorUnknownPackage
-  | PackageVersionInformationErrorNoPackageJSONLock;
 
 export interface IMatchPredicate {
   (frameworkDirectory: string): boolean;
@@ -189,163 +118,37 @@ export interface IMatchPredicate {
 
 const matchAll: IMatchPredicate = (frameworkDirectory: string) => true;
 
-async function loadFrameworkInfo(pathInFrameworksDir: string): Promise<FrameworkVersionInformation> {
-  let keyedType: KeyedType;
-  let directory: string;
-  if (pathInFrameworksDir.startsWith("keyed")) {
-    keyedType = "keyed";
-    directory = pathInFrameworksDir.substring(6);
-  } else if (pathInFrameworksDir.startsWith("non-keyed")) {
-    keyedType = "non-keyed";
-    directory = pathInFrameworksDir.substring(10);
-  } else {
-    throw "pathInFrameworksDir must start with keyed or non-keyed, but is " + pathInFrameworksDir;
-  }
-  const frameworkPath = path.resolve("..", "frameworks", pathInFrameworksDir);
-  const packageJSONPath = path.resolve(frameworkPath, "package.json");
-
-  const distDir = path.resolve(frameworkPath, "dist");
-  if (fs.existsSync(distDir)) {
-    const files = fs.readdirSync(distDir);
-    const gzFiles = files.filter((file) => path.extname(file).toLowerCase() === ".gz");
-    if (gzFiles.length > 0) {
-      return new FrameworkVersionInformationError(keyedType, directory, `Found gzipped files in "${pathInFrameworksDir}/dist".`);
-    }
-  }
-
-  if (fs.existsSync(packageJSONPath)) {
-    let packageJSON = JSON.parse(fs.readFileSync(packageJSONPath, "utf8"));
-    if (packageJSON["js-framework-benchmark"]) {
-      if (packageJSON["js-framework-benchmark"]["frameworkVersionFromPackage"]) {
-        return new FrameworkVersionInformationDynamic(
-          keyedType,
-          directory,
-          packageJSON["js-framework-benchmark"]["frameworkVersionFromPackage"].split(":"),
-          packageJSON["js-framework-benchmark"]["customURL"],
-          packageJSON["js-framework-benchmark"]["useShadowRoot"],
-          packageJSON["js-framework-benchmark"]["useRowShadowRoot"],
-          packageJSON["js-framework-benchmark"]["shadowRootName"] ?? "main-element",
-          packageJSON["js-framework-benchmark"]["buttonsInShadowRoot"] ?? true,
-          packageJSON["js-framework-benchmark"]["issues"]
-        );
-      } else if (typeof packageJSON["js-framework-benchmark"]["frameworkVersion"] === "string") {
-        return new FrameworkVersionInformationStatic(
-          keyedType,
-          directory,
-          packageJSON["js-framework-benchmark"]["frameworkVersion"],
-          packageJSON["js-framework-benchmark"]["customURL"],
-          packageJSON["js-framework-benchmark"]["useShadowRoot"],
-          packageJSON["js-framework-benchmark"]["useRowShadowRoot"],
-          packageJSON["js-framework-benchmark"]["shadowRootName"] ?? "main-element",
-          packageJSON["js-framework-benchmark"]["buttonsInShadowRoot"] ?? true,
-          packageJSON["js-framework-benchmark"]["issues"]
-        );
-      } else {
-        return new FrameworkVersionInformationError(
-          keyedType,
-          directory,
-          "package.json must contain a 'frameworkVersionFromPackage' or 'frameworkVersion' in the 'js-framework-benchmark'.property"
-        );
-      }
-    } else {
-      return new FrameworkVersionInformationError(keyedType, directory, "package.json must contain a 'js-framework-benchmark' property");
-    }
-  } else {
-    return new FrameworkVersionInformationError(keyedType, directory, "No package.json found");
-  }
-}
-
-export async function loadFrameworkVersionInformation(matchPredicate: IMatchPredicate = matchAll): Promise<FrameworkVersionInformation[]> {
-  let results = new Array<Promise<FrameworkVersionInformation>>();
-  let frameworksPath = path.resolve("..", "frameworks");
-  ["keyed", "non-keyed"].forEach((keyedType: KeyedType) => {
-    let directories = fs.readdirSync(path.resolve(frameworksPath, keyedType));
-
-    for (let directory of directories) {
-      let pathInFrameworksDir = keyedType + "/" + directory;
-      if (matchPredicate(pathInFrameworksDir)) {
-        let fi = loadFrameworkInfo(pathInFrameworksDir);
-        if (fi != null) results.push(fi);
-      }
-    }
-  });
-  return await Promise.all(results);
-}
-
-export class PackageVersionInformationResult {
-  public versions: Array<PackageVersionInformation> = [];
-  constructor(public framework: FrameworkVersionInformationDynamic) {}
-  public add(packageVersionInformation: PackageVersionInformation) {
-    this.versions.push(packageVersionInformation);
-  }
-  public getVersionName(): string {
-    if (this.versions.filter((pi) => pi instanceof PackageVersionInformationErrorNoPackageJSONLock).length > 0) {
-      return "invalid (no package-lock)";
-    }
-    return this.versions.map((version) => (version instanceof PackageVersionInformationValid ? version.version : "invalid")).join(" + ");
-  }
-  getFrameworkData(): FrameworkData {
-    return {
-      name: this.framework.directory,
-      fullNameWithKeyedAndVersion: this.framework.directory + "-v" + this.getVersionName() + "-" + this.framework.keyedType,
-      uri: this.framework.url,
-      keyed: this.framework.keyedType === "keyed",
-      useShadowRoot: this.framework.useShadowRoot,
-      useRowShadowRoot: this.framework.useRowShadowRoot,
-      shadowRootName: this.framework.shadowRootName,
-      buttonsInShadowRoot: this.framework.buttonsInShadowRoot,
-      issues: this.framework.issues,
-    };
-  }
-}
-
-export async function determineInstalledVersions(framework: FrameworkVersionInformationDynamic): Promise<PackageVersionInformationResult> {
-  let versions = new PackageVersionInformationResult(framework);
-  try {
-    console.log(`http://localhost:${config.PORT}/frameworks/${framework.keyedType}/${framework.directory}/package-lock.json`);
-    let packageLock: any = (
-      await axios.get(`http://localhost:${config.PORT}/frameworks/${framework.keyedType}/${framework.directory}/package-lock.json`)
-    ).data;
-    for (let packageName of framework.packageNames) {
-      if (packageLock.dependencies[packageName]) {
-        versions.add(new PackageVersionInformationValid(packageName, packageLock.dependencies[packageName].version));
-      } else {
-        versions.add(new PackageVersionInformationErrorUnknownPackage(packageName));
-      }
-    }
-  } catch (err) {
-    if (err.errno === "ECONNREFUSED") {
-      console.log("Can't load package-lock.json via http. Make sure the http-server is running on port 8080");
-      throw "Can't load package-lock.json via http. Make sure the http-server is running on port 8080";
-    } else if (err.response && err.response.status === 404) {
-      console.log(`package-lock.json not found for ${framework.keyedType}/${framework.directory}`);
-      versions.add(new PackageVersionInformationErrorNoPackageJSONLock());
-    } else {
-      console.log("err", err);
-      versions.add(new PackageVersionInformationErrorNoPackageJSONLock());
-    }
-  }
-  return versions;
-}
-
 export async function initializeFrameworks(matchPredicate: IMatchPredicate = matchAll): Promise<FrameworkData[]> {
-  let frameworkVersionInformations = await loadFrameworkVersionInformation(matchPredicate);
+  let lsResult ;
+  try {
+    lsResult = (
+      // FIXME https://github.com/axios/axios/issues/5008
+    await (axios as any).get(`http://${config.HOST}:${config.PORT}/ls`)
+  ).data;
+  } catch (err) {
+    console.log(err);
+    console.log(`ERROR loading frameworks from http://${config.HOST}:${config.PORT}/ls. Is the server running?`);
+    throw new Error(`ERROR loading frameworks from http://${config.HOST}:${config.PORT}/ls. Is the server running?`);
+  }
 
   let frameworks: FrameworkData[] = [];
-  for (let frameworkVersionInformation of frameworkVersionInformations) {
-    if (frameworkVersionInformation instanceof FrameworkVersionInformationDynamic) {
-      frameworks.push((await determineInstalledVersions(frameworkVersionInformation)).getFrameworkData());
-    } else if (frameworkVersionInformation instanceof FrameworkVersionInformationStatic) {
-      frameworks.push(frameworkVersionInformation.getFrameworkData());
-    } else {
-      console.log(
-        `WARNING: Ignoring package ${frameworkVersionInformation.keyedType}/${frameworkVersionInformation.directory}: ${frameworkVersionInformation.error}`
-      );
-      frameworks.push(null);
-    }
+  for (let ls of lsResult) {
+    let frameworkVersionInformation: FrameworkInformation = ls;
+    let fullName = frameworkVersionInformation.type + "/" + frameworkVersionInformation.directory;
+    if (matchPredicate(fullName)) {
+      frameworks.push({
+          name: frameworkVersionInformation.directory,
+          fullNameWithKeyedAndVersion: frameworkVersionInformation.frameworkVersionString,
+          uri: "frameworks/" + fullName + (frameworkVersionInformation.customURL ? frameworkVersionInformation.customURL : ""),
+          keyed: frameworkVersionInformation.type === "keyed",
+          useShadowRoot: !!frameworkVersionInformation.useShadowRoot,
+          useRowShadowRoot: !!frameworkVersionInformation.useRowShadowRoot,
+          shadowRootName: frameworkVersionInformation.shadowRootName,
+          buttonsInShadowRoot: !!frameworkVersionInformation.buttonsInShadowRoot,
+          issues: (frameworkVersionInformation.issues ?? []).map(i => Number(i))
+        });
+      }
   }
-
-  frameworks = frameworks.filter((f) => f !== null);
   if (config.LOG_DETAILS) {
     console.log("All available frameworks: ");
     console.log(frameworks.map((fd) => fd.fullNameWithKeyedAndVersion));
