@@ -1,27 +1,10 @@
 import * as yargs from "yargs";
-import {
-  buildDriver,
-  setUseShadowRoot,
-  testTextContains,
-  testTextNotContained,
-  testClassContains,
-  testElementLocatedByXpath,
-  testElementNotLocatedByXPath,
-  testElementLocatedById,
-  clickElementById,
-  clickElementByXPath,
-  getTextByXPath,
-  mainRoot,
-  findByXPath,
-  setUseRowShadowRoot,
-  setShadowRootName,
-  setButtonsInShadowRoot,
-} from "./webdriverAccess";
+import { checkElementContainsText, checkElementExists, clickElement, startBrowser } from "./playwrightAccess";
 import { config, FrameworkData, initializeFrameworks, BenchmarkOptions } from "./common";
-import { WebDriver, By, WebElement, logging } from "selenium-webdriver";
 
 import * as R from "ramda";
 import { valid } from "semver";
+import { ElementHandle, Page } from "playwright";
 
 let args: any = yargs(process.argv)
   .usage("$0 [--framework Framework1 Framework2 ...] [--benchmark Benchmark1 Benchmark2 ...]")
@@ -35,9 +18,6 @@ let args: any = yargs(process.argv)
 let allArgs = args._.length <= 2 ? [] : args._.slice(2, args._.length);
 
 console.log("args.framework", args.framework, !args.framework);
-
-// necessary to launch without specifiying a path
-var chromedriver: any = require("chromedriver");
 
 let init = (shadowRootName: string) => `
 window.nonKeyedDetector_reset = function() {
@@ -126,9 +106,9 @@ window.nonKeyedDetector_reset();
 function isKeyedRun(result: any, shouldBeKeyed: boolean): boolean {
   let r = result.tradded >= 1000 && result.trremoved >= 1000;
   if (r && !shouldBeKeyed) {
-    console.log(
-      `Non-keyed test for create rows failed. Expected that TRs should be recycled, but there were ${result.tradded} added TRs and ${result.trremoved} were removed`
-    );
+    // console.log(
+    //   `Non-keyed test for create rows failed. Expected that TRs should be recycled, but there were ${result.tradded} added TRs and ${result.trremoved} were removed`
+    // );
   } else if (!r && shouldBeKeyed) {
     console.log(
       `Keyed test for create rows failed. Expected that 1000 TRs should be removed and added, but there were ${result.tradded} added TRs and ${result.trremoved} were removed`
@@ -139,19 +119,18 @@ function isKeyedRun(result: any, shouldBeKeyed: boolean): boolean {
 function isKeyedRemove(result: any, shouldBeKeyed: boolean): boolean {
   let r = result.removedStoredTr;
   if (r && !shouldBeKeyed) {
-    console.log(`Note: Non-keyed test for remove is acutally keyed. Expected that the dom node for the 2nd row would NOT be removed, but it was.`);
+    // console.log(`Note: Non-keyed test for remove is acutally keyed. Expected that the dom node for the 2nd row would NOT be removed, but it was.`);
   } else if (!r && shouldBeKeyed) {
     console.log(`Keyed test for remove failed. Expected that the dom node for the 2nd row would be removed, but it wasn't`);
   }
   return r;
 }
 function isKeyedSwapRow(result: any, shouldBeKeyed: boolean): boolean {
-  //  console.log("isKeyedSwapRow", result);
-  let r = result.tradded > 0 && result.trremoved > 0 && (!shouldBeKeyed || result.newNodes == 0);
+  let r = result.tradded > 0 && result.trremoved > 0 && result.newNodes == 0;
   if (r && !shouldBeKeyed) {
-    console.log(
-      `Non-keyed test for swap failed. Expected than no TRs are added or removed, but there were ${result.tradded} added and ${result.trremoved} removed`
-    );
+    // console.log(
+    //   `Non-keyed test for swap failed. Expected than no TRs are added or removed, but there were ${result.tradded} added and ${result.trremoved} removed`
+    // );
   } else if (!r && shouldBeKeyed) {
     if (result.newNodes > 0) {
       console.log(`Keyed test for swap failed. Swap must add the TRs that it removed, but there were ${result.newNodes} new nodes`);
@@ -164,10 +143,11 @@ function isKeyedSwapRow(result: any, shouldBeKeyed: boolean): boolean {
   return r;
 }
 
-async function assertChildNodes(elem: WebElement, expectedNodes: string[], message: string) {
-  let elements = await elem.findElements(By.css("*"));
-  let allNodes = await Promise.all(elements.map((e) => e.getTagName()));
-  if (!R.equals(allNodes, expectedNodes)) {
+async function assertChildNodes(elem: ElementHandle<HTMLElement>, expectedNodes: string[], message: string) {
+  let elements = await elem.$$("*");
+  let allNodes = await Promise.all(elements.map((e) => e.evaluate(e => e.tagName)));
+  let toLower = (array: string[]) => array.map(s => s.toLowerCase());
+  if (!R.equals(toLower(allNodes), toLower(expectedNodes))) {
     console.log("ERROR in html structure for " + message);
     console.log("  expected:", expectedNodes);
     console.log("  actual  :", allNodes);
@@ -182,8 +162,8 @@ function niceEmptyString(val: string[]): string {
   return val.toString();
 }
 
-async function assertClassesContained(elem: WebElement, expectedClassNames: string[], message: string) {
-  let actualClassNames = (await elem.getAttribute("class")).split(" ");
+async function assertClassesContained(elem: ElementHandle<HTMLElement>, expectedClassNames: string[], message: string) {
+  let actualClassNames = (await elem.evaluate(e => e.className)).split(" ");
   if (!expectedClassNames.every((expected) => actualClassNames.includes(expected))) {
     console.log(
       "css class not correct. Expected for " + message + " to be " + expectedClassNames + " but was " + niceEmptyString(actualClassNames)
@@ -193,53 +173,50 @@ async function assertClassesContained(elem: WebElement, expectedClassNames: stri
   return true;
 }
 
-export async function checkTRcorrect(driver: WebDriver, timeout = config.TIMEOUT): Promise<boolean> {
-  let tr = await findByXPath(driver, "//tbody/tr[1000]", false);
-  if (!(await assertChildNodes(tr, ["td", "td", "a", "td", "a", "span", "td"], "tr"))) {
+export async function checkTRcorrect(page: Page): Promise<boolean> {
+  let tr = await page.$("tbody>tr:nth-of-type(1000)");
+  if (!(await assertChildNodes(tr as ElementHandle<HTMLElement>, ["td", "td", "a", "td", "a", "span", "td"], "tr"))) {
     return false;
   }
 
   // first td
-  let td1 = await findByXPath(driver, "//tbody/tr[1000]/td[1]", false);
-  if (!(await assertClassesContained(td1, ["col-md-1"], "first td"))) {
+  let td1 = await page.$("tbody>tr:nth-of-type(1000)>td:nth-of-type(1)");
+  if (!(await assertClassesContained(td1 as ElementHandle<HTMLElement>, ["col-md-1"], "first td"))) {
     return false;
   }
 
   // second td
-  let td2 = await findByXPath(driver, "//tbody/tr[1000]/td[2]", false);
-  if (!(await assertClassesContained(td2, ["col-md-4"], "second td"))) {
+  let td2 = await page.$("tbody>tr:nth-of-type(1000)>td:nth-of-type(2)");
+  if (!(await assertClassesContained(td2 as ElementHandle<HTMLElement>, ["col-md-4"], "second td"))) {
     return false;
   }
 
   // third td
-  let td3 = await findByXPath(driver, "//tbody/tr[1000]/td[3]", false);
-  if (!(await assertClassesContained(td3, ["col-md-1"], "third td"))) {
+  let td3 = await page.$("tbody>tr:nth-of-type(1000)>td:nth-of-type(3)");
+  if (!(await assertClassesContained(td3 as ElementHandle<HTMLElement>, ["col-md-1"], "third td"))) {
     return false;
   }
 
   // span in third td
-  let span = await findByXPath(driver, "//tbody/tr[1000]/td[3]/a/span", false);
-  if (!(await assertClassesContained(span, ["glyphicon", "glyphicon-remove"], "span in a in third td"))) {
+  let span = await page.$("tbody>tr:nth-of-type(1000)>td:nth-of-type(3)>a>span");
+  if (!(await assertClassesContained(span as ElementHandle<HTMLElement>, ["glyphicon", "glyphicon-remove"], "span in a in third td"))) {
     return false;
   }
-  let spanAria = await span.getAttribute("aria-hidden");
-  if ("true" != spanAria) {
+  // console.log("names", await span.evaluate(e => e.getAttributeNames()));
+  let spanAria = await span.evaluate(e => e.getAttribute("aria-hidden"));
+  // console.log("aria ", spanAria);
+  if ("true" !== spanAria) {
     console.log("Expected to find 'aria-hidden'=true on span in third td, but found ", spanAria);
     return false;
   }
 
-  // fourth td
-  let td4 = await findByXPath(driver, "//tbody/tr[1000]/td[4]", false);
-  if (!(await assertClassesContained(td4, ["col-md-6"], "fourth td"))) {
+  // // fourth td
+  let td4 = await page.$("tbody>tr:nth-of-type(1000)>td:nth-of-type(4)");
+  if (!(await assertClassesContained(td4 as ElementHandle<HTMLElement>, ["col-md-6"], "fourth td"))) {
     return false;
   }
 
   return true;
-}
-
-export async function getInnerHTML(driver: WebDriver, xpath: string, timeout = config.TIMEOUT): Promise<string> {
-  let elem = await findByXPath(driver, xpath, false);
-  return elem.getAttribute("innerHTML");
 }
 
 async function runBench(frameworkNames: string[]) {
@@ -248,58 +225,57 @@ async function runBench(frameworkNames: string[]) {
   runFrameworks = await initializeFrameworks(matchesDirectoryArg);
   console.log("Frameworks that will be checked", runFrameworks.map((f) => f.fullNameWithKeyedAndVersion).join(" "));
 
-  let frameworkMap = new Map<String, FrameworkData>();
+  let frameworkMap = new Map<string, FrameworkData>();
 
   let allCorrect = true;
 
+  console.log("*** headless", benchmarkOptions.headless)
+
   for (let i = 0; i < runFrameworks.length; i++) {
-    let driver = await buildDriver(benchmarkOptions);
+    let browser = await startBrowser(benchmarkOptions);
+    let page = await browser.newPage();
     try {
       let framework: FrameworkData = runFrameworks[i];
-      setUseShadowRoot(framework.useShadowRoot);
-      setUseRowShadowRoot(framework.useRowShadowRoot);
-      setShadowRootName(framework.shadowRootName);
-      setButtonsInShadowRoot(framework.buttonsInShadowRoot);
 
-      await driver.get(`http://localhost:${config.PORT}/${framework.uri}/index.html`);
-      await testElementLocatedById(driver, "add", config.TIMEOUT, true);
-      await clickElementById(driver, "run", true);
-      await testTextContains(driver, "//tbody/tr[1000]/td[1]", "1000", config.TIMEOUT, false);
+      await page.goto(`http://${config.HOST}:${config.PORT}/${framework.uri}/index.html`, {waitUntil: "networkidle"});
+      await checkElementExists(page, "#add");
+      await clickElement(page, "#add");
+      await checkElementContainsText(page, "tbody>tr:nth-of-type(1000)>td:nth-of-type(1)", "1000");
 
       // check html for tr
-      let htmlCorrect = await checkTRcorrect(driver);
+      let htmlCorrect = await checkTRcorrect(page);
       if (!htmlCorrect) {
         console.log("ERROR: Framework " + framework.fullNameWithKeyedAndVersion + " html is not correct");
         allCorrect = false;
       }
-
-      await driver.executeScript(init(framework.shadowRootName));
+      let str = init(framework.shadowRootName);
+      await page.evaluate(str);
       if (framework.useShadowRoot) {
-        await driver.executeScript(`window.nonKeyedDetector_setUseShadowDom("${framework.shadowRootName}");`);
+        await page.evaluate(`window.nonKeyedDetector_setUseShadowDom("${framework.shadowRootName}");`);
       } else {
-        await driver.executeScript(`window.nonKeyedDetector_setUseShadowDom(undefined);`);
+        await page.evaluate(`window.nonKeyedDetector_setUseShadowDom(undefined);`);
       }
-      await driver.executeScript("window.nonKeyedDetector_instrument()");
+      await page.evaluate("window.nonKeyedDetector_instrument()");
       // swap
-      await driver.executeScript("nonKeyedDetector_storeTr()");
-      await clickElementById(driver, "swaprows", true);
-      await testTextContains(driver, "//tbody/tr[2]/td[1]", "999", config.TIMEOUT, false);
-      let res = await driver.executeScript("return nonKeyedDetector_result()");
+      await page.evaluate("nonKeyedDetector_storeTr()");
+      await clickElement(page, "#swaprows");
+      await checkElementContainsText(page, "tbody>tr:nth-of-type(2)>td:nth-of-type(1)", "999");
+      let res = await page.evaluate("nonKeyedDetector_result()");
       let keyedSwap = isKeyedSwapRow(res, framework.keyed);
       // run
-      await driver.executeScript("nonKeyedDetector_storeTr()");
-      await driver.executeScript("window.nonKeyedDetector_reset()");
-      await clickElementById(driver, "run", true);
-      await testTextContains(driver, "//tbody/tr[1000]/td[1]", "2000", config.TIMEOUT, false);
-      res = await driver.executeScript("return nonKeyedDetector_result()");
+      await page.evaluate("nonKeyedDetector_storeTr()");
+      await page.evaluate("window.nonKeyedDetector_reset()");
+      await clickElement(page, "#run");
+      await checkElementContainsText(page, "tbody>tr:nth-of-type(1000)>td:nth-of-type(1)", "2000");
+      res = await page.evaluate("nonKeyedDetector_result()");
       let keyedRun = isKeyedRun(res, framework.keyed);
-      // remove
-      await driver.executeScript("nonKeyedDetector_storeTr()");
-      let text = await getTextByXPath(driver, `//tbody/tr[2]/td[2]/a`, false);
-      await driver.executeScript("window.nonKeyedDetector_reset()");
-      await clickElementByXPath(driver, `//tbody/tr[2]/td[3]/a/span[1]`, false);
-      await testTextNotContained(driver, `//tbody/tr[2]/td[2]/a`, text, config.TIMEOUT, false);
-      res = await driver.executeScript("return nonKeyedDetector_result()");
+      // // remove
+      await page.evaluate("nonKeyedDetector_storeTr()");
+      await page.evaluate("window.nonKeyedDetector_reset()");
+      await checkElementContainsText(page, `tbody>tr:nth-of-type(2)>td:nth-of-type(1)`, "1002");
+      await clickElement(page, `tbody>tr:nth-of-type(2)>td:nth-of-type(3)>a>span:nth-of-type(1)`);
+      await checkElementContainsText(page, `tbody>tr:nth-of-type(2)>td:nth-of-type(1)`, "1003");
+      res = await page.evaluate("nonKeyedDetector_result()");
       let keyedRemove = isKeyedRemove(res, framework.keyed);
       let keyed = keyedRemove && keyedRun && keyedSwap;
       console.log(
@@ -324,7 +300,8 @@ async function runBench(frameworkNames: string[]) {
       allCorrect = false;
     } finally {
       try {
-        await driver.quit();
+        await page.close();
+        await browser.close();
       } catch (e) {
         console.log("error calling driver.quit - ignoring this exception");
       }
@@ -335,9 +312,15 @@ async function runBench(frameworkNames: string[]) {
 
 config.PORT = Number(args.port);
 
+if (process.env.HOST) {
+  config.HOST = process.env.HOST;
+  console.log(`INFO: Using host ${config.HOST} instead of localhost`);
+}
+
 let runFrameworks = (args.framework && args.framework.length > 0 ? args.framework : [""]).map((v: string) => v.toString());
 
 let benchmarkOptions: BenchmarkOptions = {
+  HOST: config.HOST,
   port: config.PORT.toFixed(),
   remoteDebuggingPort: config.REMOTE_DEBUGGING_PORT,
   chromePort: config.CHROME_PORT,
@@ -352,8 +335,8 @@ async function main() {
   if (args.help) {
     yargs.showHelp();
   } else {
-    runBench(runFrameworks);
+    await runBench(runFrameworks);
   }
 }
 
-main();
+main().catch(err => {console.log("Error in isKeyed", err)});
